requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-mssql-backend==2.8.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.18.0
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mssql-django==1.7
numpy==1.24.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
Pillow==10.1.0
pipenv==2025.0.4
platformdirs==4.3.8
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.10.1
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
scikit-learn==1.3.2
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
starlette==0.27.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
virtualenv==20.32.0
Werkzeug==3.1.3
whitenoise==6.9.0
zipp==3.23.0
